[test_nn] split parametrization test from test_nn
